better

+++ src/Path.java
@@ -3,19 +3,19 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class Path {
-    Map<Integer, Integer> states = Collections.synchronizedMap(new HashMap<Integer, Integer>());
-    Map<Integer, Integer> ways = Collections.synchronizedMap(new HashMap<Integer, Integer>());
+    Map<Long, Long> states = Collections.synchronizedMap(new HashMap<Long, Long>());
+    Map<Long, Integer> ways = Collections.synchronizedMap(new HashMap<Long, Integer>());
 
-    synchronized public void put(int hash, int parentHash, int way) {
+    synchronized public void put(long hash, long parentHash, int way) {
         states.put(hash, parentHash);
         ways.put(hash, way);
     }
 
-    synchronized public int getWay(int hash) {
+    synchronized public int getWay(long hash) {
         return ways.get(hash);
     }
 
-    synchronized public int getParent(int hash) {
+    synchronized public long getParent(long hash) {
         return states.get(hash);
     }
